backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
supabase==2.0.0
asyncpg==0.29.0
sqlalchemy==2.0.23

# AI Services
openai==1.3.7
anthropic==0.7.8

# Social Media APIs
tweepy==4.14.0
facebook-sdk==3.1.0
python-linkedin-v2==2.0.1

# Data Processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
aiohttp==3.9.1

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx>=0.24.0,<0.25.0
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0

